refractor: extract workbox config generation
